refactor: hero/landing page file structure
Moved all of the components used in the hero page from the components directory to the route folder. Adjusted 'in development' disclaimer
