+++ src/app/(hero)/HeroCard.tsx
