+++ src/app/(hero)/page.tsx
@@ -1,14 +1,14 @@
-import GamemodeBlock from '@/components/hero/GamemodeBlock'
-import GamemodesWrapper from '@/components/hero/GamemodesWrapper'
-import Section from '@/components/hero/Section'
+import GamemodeBlock from './GamemodeBlock'
+import GamemodesWrapper from './GamemodesWrapper'
+import Section from '@/app/(hero)/Section'
 import { Button } from '@/components/ui/Button'
 import Link from 'next/link'
 import OneQuestionIcon from '@/public/one-question-icon.svg'
 import ExamIcon from '@/public/exam-icon.svg'
 import FlashcardsIcon from '@/public/flashcards-icon.svg'
 import SqlTrainingIcon from '@/public/sql-training-icon.svg'
-import HeroSearchBar from '@/components/hero/HeroSearchBar'
-import HeroCard from '@/components/hero/HeroCard'
+import HeroSearchBar from './HeroSearchBar'
+import HeroCard from './HeroCard'
 import { BarChart3, GalleryHorizontalEnd, History, SkullIcon, Users } from 'lucide-react'
 export default function Home() {
   return (
@@ -32,7 +32,7 @@ export default function Home() {
           </div>
         </div>
       </section>
-      <p className="py-8 text-lg font-bold text-red-500 text-center">
+      <p className="m-4 p-4 text-lg font-bold border border-red-800 bg-red-600 text-white text-center">
         <span className="uppercase">WAŻNE: </span>Strona jest publicznie dostępna w ramach
         testowania, ale wciąż jest rozwijana i nie jest w finałowym stanie. Istnieje
         możliwość resetu danych użytkowników.
